Remove 'MSVC' compiler handling
As a pure Linux developer it is hard to support other operating systems
or other compiler which are not available under Linux. Because of that
I can not guarantee the functionality of the provided source code.

To minimize maintenance and provide only functional code the 'MSVC'
compiler is no longer supported.