Rewrite function which checks compiler flags
Instead of cache-by-option the functions are now cache-by-default.
This should improve the usability of the functions and avoids
compiler flag checks which improve the CMake configure step on
CMake code changes.

The previous options was renamed and should now be used to force
a cache invalidation and a new compiler flag check.